requirements.txt
# Existing dependencies...
Flask==3.0.0
Flask-SocketIO==5.3.0
python-dotenv==1.0.0

# NEW: News Screener dependencies
fyers-apiv3==3.1.5
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
pytz==2024.1